requirements.txt
django==1.11.7
Pillow==4.3.0
djangorestframework==3.7.3